tool-go-invoice/requirements.txt
requests==2.32.5
openpyxl==3.1.5
PyQt5>=5.15.11
PyQt5-Qt5>=5.15.2
Pillow==10.2.0
playwright>=1.40.0
urllib3==2.5.0
